+++ GladiatorArenaServer/src/main/java/com/github/malow/gladiatorarena/server/game/socketnetwork/Client.java
@@ -2,13 +2,13 @@ package com.github.malow.gladiatorarena.server.game.socketnetwork;
 
 import java.net.Socket;
 
-import com.github.malow.malowlib.ProcessEvent;
+import com.github.malow.malowlib.malowprocess.ProcessEvent;
 import com.github.malow.malowlib.network.NetworkChannel;
 
 public class Client extends NetworkChannel
 {
   public boolean ready = false;
-  public Long accId;
+  public Integer accId;
 
   public boolean disconnected = false;
